feat: Import locales from JSON files dynamically (#75)
BREAKING CHANGE: The `locale` prop is now a string and fallbacks to `en-US` if it has an invalid value
