--- src/locales/de-DE.js
@@ -1,31 +1,0 @@
-export default {
-    todayButton: 'Heute',
-    nextMonth: 'Nächster Monat',
-    previousMonth: 'Letzter Monat',
-    nextYear: 'Nächstes Jahr',
-    previousYear: 'Letztes Jahr',
-    weekdays: [
-      'Sonntag',
-      'Montag',
-      'Dienstag',
-      'Mittwoch',
-      'Donnerstag',
-      'Freitag',
-      'Samstag',
-    ],
-    months: [
-      'Januar',
-      'Februar',
-      'März',
-      'April',
-      'Mai',
-      'Juni',
-      'Juli',
-      'August',
-      'September',
-      'Oktober',
-      'November',
-      'Dezember',
-    ],
-  };
-
