+++ src/locales/de-DE.json
@@ -1,0 +1,30 @@
+{
+  "todayButton": "Heute",
+  "nextMonth": "Nächster Monat",
+  "previousMonth": "Letzter Monat",
+  "nextYear": "Nächstes Jahr",
+  "previousYear": "Letztes Jahr",
+  "weekdays": [
+    "Sonntag",
+    "Montag",
+    "Dienstag",
+    "Mittwoch",
+    "Donnerstag",
+    "Freitag",
+    "Samstag"
+  ],
+  "months": [
+    "Januar",
+    "Februar",
+    "März",
+    "April",
+    "Mai",
+    "Juni",
+    "Juli",
+    "August",
+    "September",
+    "Oktober",
+    "November",
+    "Dezember"
+  ]
+}
