--- src/locales/es-ES.js
@@ -1,30 +1,0 @@
-export default {
-  todayButton: 'Hoy',
-  nextMonth: 'Próximo mes',
-  previousMonth: 'Mes anterior',
-  nextYear: 'Próximo año',
-  previousYear: 'Año anterior',
-  weekdays: [
-    'Domingo',
-    'Lunes',
-    'Martes',
-    'Miércoles',
-    'Jueves',
-    'Viernes',
-    'Sábado',
-  ],
-  months: [
-    'Enero',
-    'Febrero',
-    'Marzo',
-    'Abril',
-    'Mayo',
-    'Junio',
-    'Julio',
-    'Agosto',
-    'Septiembre',
-    'Octubre',
-    'Noviembre',
-    'Diciembre',
-  ],
-};
