--- src/locales/it-IT.js
@@ -1,30 +1,0 @@
-export default {
-  todayButton: 'Oggi',
-  nextMonth: 'Prossimo mese',
-  previousMonth: 'Mese scorso',
-  nextYear: "L'anno prossimo",
-  previousYear: "L'anno scorso",
-  weekdays: [
-    'Domenica',
-    'Lunedì',
-    'Martedì',
-    'Mercoledì',
-    'Giovedì',
-    'Venerdì',
-    'Sabato',
-  ],
-  months: [
-    'Gennaio',
-    'Febbraio',
-    'Marzo',
-    'Aprile',
-    'Maggio',
-    'Giugno',
-    'Luglio',
-    'Agosto',
-    'Settembre',
-    'Ottobre',
-    'Novembre',
-    'Dicembre',
-  ],
-};
