+++ src/locales/it-IT.json
@@ -1,0 +1,30 @@
+{
+  "todayButton": "Oggi",
+  "nextMonth": "Prossimo mese",
+  "previousMonth": "Mese scorso",
+  "nextYear": "L'anno prossimo",
+  "previousYear": "L'anno scorso",
+  "weekdays": [
+    "Domenica",
+    "Lunedì",
+    "Martedì",
+    "Mercoledì",
+    "Giovedì",
+    "Venerdì",
+    "Sabato"
+  ],
+  "months": [
+    "Gennaio",
+    "Febbraio",
+    "Marzo",
+    "Aprile",
+    "Maggio",
+    "Giugno",
+    "Luglio",
+    "Agosto",
+    "Settembre",
+    "Ottobre",
+    "Novembre",
+    "Dicembre"
+  ]
+}
